Assignment 4 - The Mechanic Shop
My attempt to make an extremally simple PoS system for a mechanic shop and my personal favorite project that whole semester. i completed it the day it was given to me just because i had so much fun sitting down and messing with all the methods and loops I could implement.